# apps/cms-scaffold/docs/v2/contracts/BW6.openapi.yaml
openapi: 3.0.3
info:
  title: CMS Scaffold API
  description: "Kernel API for CMS Scaffold. Handwritten OpenAPI is the contract source.\n\nPublic reads live under\
    \ `/api/v1/public/**`. Work surfaces do not share\nthat prefix. Session cookie name is `cms_session` (one cookie).\n\
    Published media is not a public static directory.\n\nSurface is resolved from the `Origin` header (see `cms.identity.surface-origins`).\n\
    A cookie request without `Origin` is treated as the Front surface. A bearer request\nwithout `Origin` may name\
    \ its surface with `X-CMS-Surface`.\n\nEvery error response uses `ErrorEnvelope`; `error.code` is one of `ErrorCode`.\n\
    Each operation lists its error codes in its description.\n\nCommon errors (not repeated per operation):\n- 401\
    \ `UNAUTHENTICATED` (no session) or `SESSION_EXPIRED` (expired or revoked session)\n  on every operation that\
    \ is not marked `security: []`.\n- 403 `CSRF_FAILED` on every POST, PUT, PATCH, DELETE sent with a `cms_session`\
    \ cookie\n  when the `X-CSRF-Token` header does not equal the `cms_csrf` cookie.\n- 400 `VALIDATION_FAILED`\
    \ when a path or query parameter has the wrong type\n  (for example a malformed UUID) or a JSON body cannot\
    \ be parsed.\n- 405 `METHOD_NOT_ALLOWED` when the path exists but the HTTP method does not.\n- 415 `MEDIA_TYPE_NOT_SUPPORTED`\
    \ when a JSON operation receives another Content-Type.\n- 500 `INTERNAL_ERROR` for any unexpected server failure.\n\
...
servers:
...
paths:
...
  /api/v1/auth/login:
    post:
      operationId: login
      tags:
      - Auth
      security: []
      description: "Errors:\n- 400 VALIDATION_FAILED: username or password missing or blank.\n- 401 INVALID_CREDENTIALS:\
        \ unknown username or wrong password.\n- 403 ACCOUNT_DISABLED: principal is disabled.\n- 403 ACCOUNT_LOCKED:\
        \ principal is locked after repeated failures.\n- 403 CSRF_FAILED: `Origin` is present and not in `cms.identity.cors-origins`,\n\
        \  or the request already carries a session cookie without a matching CSRF token.\n"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated; sets cms_session and cms_csrf cookies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/auth/logout:
    post:
      operationId: logout
      tags:
      - Auth
      parameters:
      - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '204':
          description: Current session revoked; cookies cleared
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/auth/me:
    get:
      operationId: me
      tags:
      - Auth
      responses:
        '200':
          description: Current principal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/auth/csrf:
    get:
      operationId: csrf
      tags:
      - Auth
      security: []
      responses:
        '200':
          description: Issues a new cms_csrf cookie and returns the same token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfToken'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/auth/password/change:
    post:
      operationId: changePassword
      tags:
      - Auth
      description: 'Errors:

        - 400 VALIDATION_FAILED: new password shorter than 12 characters or equal to the username.

        - 401 INVALID_CREDENTIALS: current password is wrong.

        '
      parameters:
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        '204':
          description: Password changed; all sessions revoked; cookies cleared
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
...
  /api/v1/roles:
    get:
      operationId: listRoles
      tags:
      - Principals
      description: 'Errors:

        - 403 SURFACE_FORBIDDEN, FORBIDDEN: as listPrincipals.

        '
      responses:
        '200':
          description: Roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleList'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'
...
components:
  parameters:
...
    CsrfHeader:
      in: header
      name: X-CSRF-Token
      required: false
      description: Required when the request carries the cms_session cookie; must equal the cms_csrf cookie.
      schema:
        type: string
  responses:
    Error400:
      description: Bad request. See the operation description for codes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error401:
      description: UNAUTHENTICATED, SESSION_EXPIRED or INVALID_CREDENTIALS.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error403:
      description: Forbidden. See the operation description for codes; CSRF_FAILED applies to every unsafe cookie
        request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Error415:
      description: Unsupported media type. See the operation description for codes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Error500:
      description: INTERNAL_ERROR.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cms_session
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ErrorCode:
      type: string
      description: 'Every `error.code` the API can return. Media codes were lowercase until BW5 (BQ-07).

        '
      enum:
      - UNAUTHENTICATED
      - INVALID_CREDENTIALS
      - SESSION_EXPIRED
      - ACCOUNT_DISABLED
      - ACCOUNT_LOCKED
      - CSRF_FAILED
      - FORBIDDEN
      - SURFACE_FORBIDDEN
      - VALIDATION_FAILED
      - LAST_ADMIN
      - ENTRY_NOT_FOUND
      - CONTENT_TYPE_NOT_FOUND
      - NAVIGATION_NOT_FOUND
      - AUDIT_EVENT_NOT_FOUND
      - PRINCIPAL_NOT_FOUND
      - AUDIENCE_PARAM_REJECTED
      - INVALID_STATE_TRANSITION
      - SLUG_CONFLICT
      - VERSION_CONFLICT
      - VERSION_REQUIRED
      - TYPE_DISABLED
      - TYPE_IN_USE
      - REF_CONSTRAINT
      - SINGLETON_EXISTS
      - SLUG_REQUIRED
      - FIELD_VALIDATION
      - REF_TARGET_NOT_FOUND
      - REF_TARGET_WRONG_TYPE
      - PRINCIPAL_REF_UNRESOLVED
      - MEDIA_NOT_FOUND
      - MEDIA_VARIANT_NOT_AVAILABLE
      - MEDIA_UNSUPPORTED_TYPE
      - MEDIA_QUOTA_EXCEEDED
      - MEDIA_FILE_TOO_LARGE
      - MEDIA_GONE
      - ROUTE_NOT_FOUND
      - METHOD_NOT_ALLOWED
      - MEDIA_TYPE_NOT_SUPPORTED
      - RATE_LIMITED
      - INTERNAL_ERROR
      - CONFIRMATION_REQUIRED
      - SELF_DISABLE_FORBIDDEN
      - SELF_DEMOTION_FORBIDDEN
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
      - error
      - requestId
      properties:
        requestId:
          type: string
          description: Echo of X-Request-Id, or a server-generated UUID.
        error:
          type: object
          additionalProperties: false
          required:
          - code
          - message
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: English developer message. Not for display.
            action:
              type: string
              description: Present on FORBIDDEN and SURFACE_FORBIDDEN.
            contentType:
              type: string
              description: Present when the denied action targets a content type.
            surface:
              type: string
              description: Present on FORBIDDEN and SURFACE_FORBIDDEN.
            fields:
              type: array
              minItems: 1
              description: Present on 422 payload, content-type or settings validation errors; every invalid key,
                in validation order.
              items:
                $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      additionalProperties: false
      required:
      - field
      - code
      - message
      properties:
        field:
          type: string
          description: Input path (`payload.<fieldKey>`), a settings property, or a content-type body path such
            as `fields[0].key`.
          example: payload.title
        code:
          $ref: '#/components/schemas/FieldErrorCode'
        message:
          type: string
          description: English developer message. Not for display.
    FieldErrorCode:
      type: string
      description: 'Why a field is invalid. REQUIRED (publish, member create, or missing content-type/settings input),
        RESERVED_KEY, WRONG_TYPE,

        TOO_LONG (the endpoint-specific Unicode code-point limit; payload string 1,000 or markdown 100,000),

        INVALID_DATETIME, NOT_IN_ENUM, INVALID_UUID, REF_TARGET_NOT_FOUND,

        REF_TARGET_WRONG_TYPE, PRINCIPAL_REF_UNRESOLVED, INVALID_FORMAT (a value that must match a pattern), DUPLICATE

        (a key that already exists or repeats in the request).

        '
      enum:
      - REQUIRED
      - RESERVED_KEY
      - WRONG_TYPE
      - TOO_LONG
      - INVALID_DATETIME
      - NOT_IN_ENUM
      - INVALID_UUID
      - REF_TARGET_NOT_FOUND
      - REF_TARGET_WRONG_TYPE
      - PRINCIPAL_REF_UNRESOLVED
      - INVALID_FORMAT
      - DUPLICATE
...
    Surface:
      type: string
      enum:
      - front
      - back
      - admin
    CmsAction:
      type: string
      enum:
      - read_published
      - read_draft
      - create
      - update
      - publish
      - unpublish
      - delete
      - archive
      - manage_media
      - manage_types
      - manage_principals
      - manage_settings
      - read_audit
    PrincipalStatus:
      type: string
      enum:
      - active
      - disabled
      - locked
...
    LoginRequest:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        surface:
          type: string
          description: Ignored; the surface comes from Origin.
    CsrfToken:
      type: object
      additionalProperties: false
      required:
      - csrfToken
      properties:
        csrfToken:
          type: string
    MePrincipal:
      type: object
      additionalProperties: false
      required:
      - id
      - username
      - displayName
      - status
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        status:
          $ref: '#/components/schemas/PrincipalStatus'
    RoleAssignment:
      type: object
      additionalProperties: false
      required:
      - code
      - contentTypeCodes
      properties:
        code:
          type: string
        contentTypeCodes:
          type: array
          items:
            type: string
    SurfaceFlags:
      type: object
      additionalProperties: false
      required:
      - front
      - back
      - admin
      properties:
        front:
          type: boolean
        back:
          type: boolean
        admin:
          type: boolean
    TypeCapability:
      type: object
      additionalProperties: false
      required:
      - key
      - actions
      - scoped
      properties:
        key:
          type: string
        actions:
          type: array
          description: Subset of read_published, read_draft, create, update, publish, unpublish, delete, archive,
            in that order.
          items:
            $ref: '#/components/schemas/CmsAction'
        scoped:
          type: boolean
          description: True when some listed action is granted only with a predicate; the caller must still expect
            403 on some entries.
    Capabilities:
      type: object
      additionalProperties: false
      required:
      - surface
      - types
      - global
      description: 'What the caller may do on the surface of this request (G-01). Only enabled types with at least
        one action

        are listed, in type key order. Surface hard-deny rules are already applied.

        '
      properties:
        surface:
          $ref: '#/components/schemas/Surface'
        types:
          type: array
          items:
            $ref: '#/components/schemas/TypeCapability'
        global:
          type: array
          description: Subset of manage_media, manage_types, manage_principals, manage_settings, read_audit, in
            that order.
          items:
            $ref: '#/components/schemas/CmsAction'
    Me:
      type: object
      additionalProperties: false
      required:
      - principal
      - roles
      - surfaces
      - capabilities
      properties:
        principal:
          $ref: '#/components/schemas/MePrincipal'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleAssignment'
        surfaces:
          $ref: '#/components/schemas/SurfaceFlags'
        capabilities:
          $ref: '#/components/schemas/Capabilities'
    LoginResponse:
      type: object
      additionalProperties: false
      required:
      - principal
      - roles
      - surfaces
      - capabilities
      - csrfToken
      properties:
        principal:
          $ref: '#/components/schemas/MePrincipal'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleAssignment'
        surfaces:
          $ref: '#/components/schemas/SurfaceFlags'
        capabilities:
          $ref: '#/components/schemas/Capabilities'
        csrfToken:
          type: string
    PasswordChangeRequest:
      type: object
      required:
      - currentPassword
      - newPassword
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 12
...
    Role:
      type: object
      additionalProperties: false
      required:
      - code
      - displayName
      - system
      properties:
        code:
          type: string
        displayName:
          type: string
        system:
          type: boolean
    RoleList:
      type: object
      additionalProperties: false
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Role'
...
security:
- cookieAuth: []
- bearerAuth: []
...